requirements.txt
keras==2.9.0
lime==0.2.0.1
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
scikit_image==0.19.2
skimage==0.0
streamlit==1.11.1
tensorflow==2.9.1
